Move rules engine to workers-shared
