+++ packages/pages-shared/asset-server/handler.ts
@@ -1,4 +1,8 @@
 import {
+	generateRulesMatcher,
+	replacer,
+} from "@cloudflare/workers-shared/asset-worker/src/utils/rules-engine";
+import {
 	FoundResponse,
 	InternalServerErrorResponse,
 	MethodNotAllowedResponse,
@@ -11,7 +15,6 @@ import {
 	SeeOtherResponse,
 	TemporaryRedirectResponse,
 } from "./responses";
-import { generateRulesMatcher, replacer } from "./rulesEngine";
 import type {
 	Metadata,
 	MetadataHeadersEntries,
