+++ packages/workers-shared/asset-worker/src/utils/rules-engine.ts
